Refactor Save/Add update method
This commit enables the the contact's status to be updated. We are
unable to modify it via the add_or_update endpoint so additional work
was required to post the soap object to the correct endpoint.
Adds the class and instance methods for update, and extracts code needed
for both save and update to a private method: prepare_api_response.

[#109653778]